REST methods added

+++ src/main/java/org/hugevictory/hugevictory/repository/StudentService.java
@@ -33,13 +33,33 @@ public class StudentService
 		return studentRepository.findById(id);
 	}
 	
-	public void updateStudent(int id, Student student)
+	public String updateStudent(int id, Student student)
 	{
-		studentRepository.save(student);
+		Optional<Student> studentOptional = studentRepository.findById(id);
+		if(!studentOptional.isPresent())
+		{
+			student.setId(id);
+			studentRepository.save(student);
+			return "student added";
+		}
+		else
+		{
+			student.setId(id);
+			studentRepository.save(student);
+			return "student updated";
+		}
 	}
 	
-	public void deleteStudent(int id)
+	public String deleteStudent(int id)
 	{
-		studentRepository.deleteById(id);
+		if(studentRepository.existsById(id))
+		{
+			studentRepository.deleteById(id);
+			return "student deleted";
+		}
+		else
+		{
+			return "student not found";
+		}
 	}
 }
